services/backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0